FIX crash as overcrowded-field was too small
git-svn-id: svn://tron.homeunix.org/simutrans/simutrans/trunk@7751 8aca7d54-2c30-db11-9de9-000461428c89

+++ simhalt.cc
@@ -2838,7 +2838,7 @@ void haltestelle_t::recalc_status()
 
 	// now for all goods
 	if(status_color!=COL_RED  &&  get_ware_enabled()) {
-		const uint32 count = warenbauer_t::get_waren_anzahl();
+		const uint8  count = warenbauer_t::get_waren_anzahl();
 		const uint32 max_ware = get_capacity(2);
 		for(  uint32 i = 3;  i < count;  i++  ) {
 			ware_besch_t const* const wtyp = warenbauer_t::get_info(i);
